fixes the json file and the naming of the server.js file
